% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-daemon.R
\docType{class}
\name{age-class}
\alias{age}
\alias{age-class}
\title{age: Age structured model daemon}
\description{
age: Age structured model daemon
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name, object.name = ".daemon", verbose = TRUE,
  autoconstruct = FALSE, conf = NULL)}}{Initialize the daemon}
}}

