% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C}
\alias{C}
\title{Catches-at-age of Gulf menhaden in the U.S. Gulf of Mexico, 1964-2004.}
\format{A data frame with 41 rows and 6 columns
\describe{
  \item{age0}{annual estimated catch at age zero}
  \item{age1}{annual estimated catch at age one}
  ...
}}
\source{
\url{http://www.sciencedirect.com/science/article/pii/S0165783606003638}
}
\usage{
C
}
\description{
A data set containing estimated annual catches-at-age, in number of fish, 
of Brevoortia patronus, from the reduction fishery. 
Extracted from Table 2 of Vaughan et al., 2007, Fish. Res. 83, 263-275.
}
\keyword{datasets}

