% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-conf.R
\docType{class}
\name{rcage.conf-class}
\alias{rcage.conf}
\alias{rcage.conf-class}
\title{Template configuration for this package}
\description{
Template configuration for this package
}
\section{Methods}{

\describe{
\item{\code{construct()}}{Constructor of virtual conf}
}}

