% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-strategy.R
\docType{class}
\name{rcage.strategy-class}
\alias{rcage.strategy}
\alias{rcage.strategy-class}
\title{strategy class}
\description{
strategy class
}
\section{Methods}{

\describe{
\item{\code{get.compute.alpha()}}{Selectivity function intercept}

\item{\code{get.compute.aM()}}{Age at maturity (aM)}

\item{\code{get.compute.aR()}}{Age at recruitment}

\item{\code{get.compute.aS(aM, aU)}}{Mean spawning age}

\item{\code{get.compute.aU()}}{Species longevity}

\item{\code{get.compute.C()}}{Catch data}

\item{\code{get.compute.chi()}}{Log-catchability}

\item{\code{get.compute.E()}}{Effort data}

\item{\code{get.compute.eta()}}{Selectivity function slope}

\item{\code{get.compute.G(nC, aM, nS, phi)}}{Transition matrix}

\item{\code{get.compute.h()}}{Initial system variance}

\item{\code{get.compute.iniM(aR, aS, aU, phi, xi, chi)}}{Initial mean of state vector}

\item{\code{get.compute.l(r, v, nC)}}{Log-likelihood}

\item{\code{get.compute.nC(aU)}}{Number of cohorts (nC)}

\item{\code{get.compute.nS(aM, aU)}}{Mean spawning age}

\item{\code{get.compute.o()}}{Measurement error variance}

\item{\code{get.compute.omega()}}{System error variance}

\item{\code{get.compute.phi()}}{Elasticity}

\item{\code{get.compute.r(t, v, G, C, E, U, s, nC, iniM, o, chi, rho, xi)}}{System mean vectors}

\item{\code{get.compute.rho()}}{Log-carrying capacity}

\item{\code{get.compute.s(aU, eta, alpha)}}{Selectivity function}

\item{\code{get.compute.t()}}{Time}

\item{\code{get.compute.theta(r, v, t, G)}}{Log-abundance}

\item{\code{get.compute.U(nC, aS, phi)}}{System intercept vector}

\item{\code{get.compute.v(t, h, o, nC, G, omega, C)}}{System variance matrices}

\item{\code{get.compute.xi()}}{Natural mortality rate}
}}

